requirement.txt
mistralai==1.6.0  
python-dotenv==1.0.1
langchain==0.3.21
requests==2.32.2
langchain-community==0.3.20
langchain_mistralai==0.3.21
langgraph==0.3.18